req.py
import torchvision
from torch.optim import lr_scheduler
from sklearn.metrics import f1_score
from sklearn.model_selection import train_test_split
from torchvision.transforms.transforms import Resize
import pickle
import numpy as np
from skimage import io

from tqdm import tqdm, tqdm_notebook
from PIL import Image
from pathlib import Path

from torchvision import transforms
from multiprocessing.pool import ThreadPool
from sklearn.preprocessing import LabelEncoder
from torch.utils.data import Dataset, DataLoader
import torch.nn as nn

from matplotlib import colors, pyplot as plt

import warnings
import pandas as pd