requirements.txt
matplotlib==3.6.2
numpy==1.23.4
Pillow==9.3.0
scikit_image==0.19.3
scikit_learn==1.1.3
skimage==0.0
torch==1.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.64.1
